Login functionality on the login page
Commented out the dashboard thing in the backend that was used for testing. Set up cors. Added login functionality to the login page that seems to be working and redirecting properly, notifying the user when the pw is wrong, and storing the user id in the session cookie upon successful login.