Frontend login functionality
Added logout functionality to the backend.

Removed the console.log(users) from the backend leaderboard bc it was clogging the logs and can be viewed in mongodb.

Fixed some server.js stuff with the cookies and simplified the cors setup.

Added an AuthContext middleware to tracker user authentication.

Added a logout button to the navbar for testing purposes. It kinda works, the page just needs to be refreshed to rerender the navbar. Will work on that soon.